refactor subfunction in utility folder
